requirements_app.txt
numpy==1.23.4
face_alignment==1.3.5
imageio==2.19.3
imageio-ffmpeg==0.4.7
librosa==0.9.2 
numba
resampy==0.3.1
pydub==0.25.1 
scipy==1.10.1
kornia==0.6.8
tqdm
yacs==0.1.8
pyyaml  
joblib==1.1.0
facexlib==0.3.0
gradio==3.38.*
#gfpgan
edge-tts
openai-whisper
scikit-image==0.19.3
zh_conv
accelerate
transformers
sentencepiece
